:arrow_up: Use non-int k with wtl::negative_binomial_distribution #4

+++ src/individual.hpp
@@ -107,7 +107,7 @@ class Individual {
     static double RECRUITMENT_COEF_;
     //! \f$k\f$ for overdispersion in recruitment()
     //! @ingroup params
-    static uint_fast32_t NEGATIVE_BINOM_K_;
+    static double NEGATIVE_BINOM_K_;
     //! mortality due to natural causes
     static std::vector<double> NATURAL_MORTALITY_;
     //! mortality due to fishing activities
